fix[2.2]: refatora caminhos de validação do servidor HTTP e asserções do Unity para o áudio

+++ Audio/main/audio_manager.h
@@ -50,7 +50,7 @@ void audio_stop(void);
 
 bool audio_is_playing(void);
 
-void audio_set_volume(int8_t volume);
+void audio_set_volume(int volume);
 
 #ifdef IS_TEST_ENVIRONMENT
     #include "audio_manager.c"
